chore(domain): move CachedTransaction to floresta-domain wallet model
Moves the shared CachedTransaction struct from floresta-watch-only to
the floresta-domain wallet model module to establish a single source of
truth for this type, enabling reuse across the electrum, node, and
watch-only crates.
