+++ crates/floresta-electrum/src/electrum_protocol.rs
@@ -22,8 +22,8 @@ use floresta_common::spsc::Channel;
 use floresta_common::try_and_log;
 use floresta_compact_filters::flat_filters_store::FlatFiltersStore;
 use floresta_compact_filters::network_filters::NetworkFilters;
+use floresta_domain::wallet::model::CachedTransaction;
 use floresta_watch_only::AddressCache;
-use floresta_watch_only::CachedTransaction;
 use floresta_watch_only::kv_database::KvDatabase;
 use floresta_wire::node_handle::NodeHandle;
 use floresta_wire::node_interface::ChainMethods;
