+++ crates/floresta-watch-only/src/lib.rs
@@ -8,7 +8,6 @@
 )]
 #![allow(clippy::manual_is_multiple_of)]
 
-use core::cmp::Ordering;
 use core::fmt::Debug;
 use std::path::Path;
 use std::sync::RwLockReadGuard;
@@ -30,14 +29,12 @@ use bitcoin::Block;
 use bitcoin::OutPoint;
 use bitcoin::Transaction;
 use bitcoin::TxOut;
-use bitcoin::consensus::deserialize;
 use bitcoin::consensus::encode::serialize_hex;
 use bitcoin::hash_types::Txid;
-use bitcoin::hashes::Hash as HashTrait;
-use bitcoin::hashes::hex::FromHex;
 use bitcoin::hashes::sha256::Hash;
 use floresta_common::prelude::*;
 use floresta_domain::wallet::error::WatchOnlyError;
+use floresta_domain::wallet::model::CachedTransaction;
 use floresta_domain::wallet::model::MerkleProof;
 use serde::Deserialize;
 use serde::Serialize;
@@ -54,49 +51,6 @@ const DERIVATION_COUNT: u32 = 100;
 
 /// Initial index for address derivation.
 const INDEX_INITIAL: u32 = 0;
-
-/// Every address contains zero or more associated transactions, this struct defines what
-/// data we store for those.
-#[derive(Debug, Clone, Eq, Serialize, Deserialize)]
-pub struct CachedTransaction {
-    pub tx: Transaction,
-    pub height: u32,
-    pub merkle_block: Option<MerkleProof>,
-    pub hash: Txid,
-    pub position: u32,
-}
-
-impl Ord for CachedTransaction {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.height.cmp(&other.height)
-    }
-}
-
-impl PartialOrd for CachedTransaction {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for CachedTransaction {
-    fn eq(&self, other: &Self) -> bool {
-        self.height == other.height
-    }
-}
-
-impl Default for CachedTransaction {
-    fn default() -> Self {
-        Self {
-            // A placeholder transaction with no input and no outputs, the bare-minimum to be
-            // serializable
-            tx: deserialize(&Vec::from_hex("010000000000ffffffff").unwrap()).unwrap(),
-            height: 0,
-            merkle_block: None,
-            hash: Txid::all_zeros(),
-            position: 0,
-        }
-    }
-}
 
 /// An address inside our cache, contains all information we need to satisfy electrum's requests
 #[derive(Debug, Clone, Serialize, Deserialize)]
